# backend/entrypoints/http/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Eggplant API
  description: |
    Eggplant is a self-hosted music streaming service .
  version: 0.0.0
servers:
  - url: /
    description: The running Eggplant instance.
...
security:
  - {}
  - cookieAuth: []
paths:
  /api/browse:
    get:
      tags: [browse]
      operationId: browse
      summary: Return the root of the music library.
      responses:
        '200':
          description: The root album.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RootAlbum'
        '503':
          $ref: '#/components/responses/LibraryNotReady'
  /api/browse/{id}:
    get:
      tags: [browse]
      operationId: browseAlbum
      summary: Return a single album by id.
      parameters:
        - $ref: '#/components/parameters/AlbumId'
      responses:
        '200':
          description: The requested album.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/LibraryNotReady'
...
  /api/search:
    get:
      tags: [browse]
      operationId: search
      summary: Search albums and tracks.
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
            example: eggplant
          description: The search query.
      responses:
        '200':
          description: Search results.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResults'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/track/{trackId}/duration:
    get:
      tags: [stream]
      operationId: getTrackDuration
      summary: Return the duration of a track.
      parameters:
        - $ref: '#/components/parameters/TrackId'
      responses:
        '200':
          description: The track duration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackDuration'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/LibraryNotReady'
  /api/track/{trackId}/stream:
    post:
      tags: [stream]
      operationId: startTrackStream
      summary: Start an HLS stream for a track.
      parameters:
        - $ref: '#/components/parameters/TrackId'
        - name: seek
          in: query
          required: false
          schema:
            type: number
            format: double
            example: 42.5
          description: Seek position in seconds.
      responses:
        '200':
          description: The created stream.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartStreamResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/track/{trackId}/stream/{streamId}/playlist:
    get:
      tags: [stream]
      operationId: getStreamPlaylist
      summary: Return the HLS playlist for a stream.
      parameters:
        - $ref: '#/components/parameters/TrackId'
        - $ref: '#/components/parameters/StreamId'
      responses:
        '200':
          description: The HLS playlist.
          content:
            application/vnd.apple.mpegurl:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequestEmpty'
        '404':
          $ref: '#/components/responses/NotFoundEmpty'
        '503':
          $ref: '#/components/responses/ServiceUnavailableEmpty'
  /api/track/{trackId}/stream/{streamId}/init:
    get:
      tags: [stream]
      operationId: getStreamInit
      summary: Return the HLS init segment for a stream.
      parameters:
        - $ref: '#/components/parameters/TrackId'
        - $ref: '#/components/parameters/StreamId'
      responses:
        '200':
          description: The init segment.
          content:
            video/mp4:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequestEmpty'
        '404':
          $ref: '#/components/responses/NotFoundEmpty'
        '503':
          $ref: '#/components/responses/ServiceUnavailableEmpty'
  /api/track/{trackId}/stream/{streamId}/fragment/{number}:
    get:
      tags: [stream]
      operationId: getStreamFragment
      summary: Return a single HLS fragment for a stream.
      parameters:
        - $ref: '#/components/parameters/TrackId'
        - $ref: '#/components/parameters/StreamId'
        - name: number
          in: path
          required: true
          schema:
            type: integer
          description: The fragment number.
      responses:
        '200':
          description: The fragment.
          content:
            video/iso.segment:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequestEmpty'
        '404':
          $ref: '#/components/responses/NotFoundEmpty'
        '503':
          $ref: '#/components/responses/ServiceUnavailableEmpty'
...
  /api/thumbnail/{id}:
    get:
      tags: [browse]
      operationId: getThumbnail
      summary: Return a thumbnail image.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The thumbnail id.
      responses:
        '200':
          description: The thumbnail image.
          content:
            image/webp:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequestEmpty'
        '404':
          $ref: '#/components/responses/NotFoundEmpty'
  /api/auth/register-initial:
    post:
      tags: [auth]
      operationId: registerInitial
      summary: Register the initial administrator account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterInitialInput'
      responses:
        '200':
          $ref: '#/components/responses/Empty'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/auth/register:
    post:
      tags: [auth]
      operationId: register
      summary: Register a new account using an invitation token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterInput'
      responses:
        '200':
          $ref: '#/components/responses/Empty'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Log in and obtain an access token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
      responses:
        '200':
          description: The access token. Also intended to be stored in the auth-token cookie.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/auth/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: Log out the current session.
      security:
        - cookieAuth: []
      responses:
        '200':
          $ref: '#/components/responses/Empty'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/create-invitation:
    post:
      tags: [auth]
      operationId: createInvitation
      summary: Create an invitation token (administrators only).
      security:
        - cookieAuth: []
      responses:
        '200':
          description: The invitation token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateInvitationResult'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/auth:
    get:
      tags: [auth]
      operationId: getCurrentUser
      summary: Return the currently authenticated user.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: The current user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadUserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/users:
    get:
      tags: [auth]
      operationId: getUsers
      summary: List all users (administrators only).
      security:
        - cookieAuth: []
      responses:
        '200':
          description: All users.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReadUserResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/auth/users/{username}/remove:
    post:
      tags: [auth]
      operationId: removeUser
      summary: Remove a user (administrators only).
      security:
        - cookieAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
          description: The username to remove.
      responses:
        '200':
          $ref: '#/components/responses/Empty'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/remote:
    get:
      tags: [remote]
      operationId: listRemotes
      summary: List paired and pending remote instances (administrators only).
      security:
        - cookieAuth: []
      responses:
        '200':
          description: The known remote instances.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RemoteInstance'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [remote]
      operationId: addRemote
      summary: Begin pairing with a peer instance (administrators only).
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddRemoteInput'
      responses:
        '200':
          description: The local pairing identifiers to hand to the peer operator.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddRemoteResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/peer/auth-token:
    post:
      tags: [remote]
      operationId: setRemoteAuthToken
      summary: Peer-to-peer endpoint a peer calls to deliver our outbound token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetRemoteAuthTokenInput'
      responses:
        '200':
          $ref: '#/components/responses/Empty'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth-token
  parameters:
    AlbumId:
      name: id
      in: path
      required: true
      schema:
        type: string
        example: 8X4F2K9MNPQRSTVW
      description: The album id.
...
  responses:
...
    LibraryNotReady:
      description: The music library is being prepared.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    RootAlbum:
      type: object
      required: [albums, tracks]
      properties:
        thumbnail:
          allOf:
            - $ref: '#/components/schemas/Thumbnail'
          nullable: true
        albums:
          type: array
          items:
            $ref: '#/components/schemas/PartialAlbum'
        tracks:
          type: array
          items:
            $ref: '#/components/schemas/Track'
    Album:
      type: object
      required: [id, title, parents, albums, tracks]
      properties:
        id:
          type: string
          example: 8X4F2K9MNPQRSTVW
        title:
          type: string
          example: Eggplant - Greatest Hits
        thumbnail:
          allOf:
            - $ref: '#/components/schemas/Thumbnail'
          nullable: true
        parents:
          type: array
          items:
            $ref: '#/components/schemas/PartialAlbum'
        albums:
          type: array
          items:
            $ref: '#/components/schemas/PartialAlbum'
        tracks:
          type: array
          items:
            $ref: '#/components/schemas/Track'
...